test(cli): split build.test.ts by concern
`build.test.ts` was 203s of the CLI suite's 317s wall clock, and jest
parallelises across files rather than within one — so no amount of workers
could shorten it. It was the tail of the entire run.

Split by concern into build (core behaviour), build-optimize, build-check-wasm
and build-check-deploy. Every test body moves verbatim; all 32 test names and
all 10 snapshots are unchanged apart from the two whose keys pick up the new
describe prefix.

Fixture ownership is now a clean partition, which the split has to preserve:
every one of these tests calls `cleanFixture`, which removes the fixture's
`.mops`, and files run in parallel workers. `--check-deploy works without a
PocketIC pin` therefore stays in build.test.ts with the rest of
`build/success` rather than moving to build-check-deploy.
