+++ cli/tests/build.test.ts
@@ -1,25 +1,17 @@
 import { describe, expect, jest, test } from "@jest/globals";
 import { execa } from "execa";
 import { createHash } from "node:crypto";
-import {
-  appendFileSync,
-  existsSync,
-  linkSync,
-  readFileSync,
-  rmSync,
-} from "node:fs";
+import { appendFileSync, existsSync, readFileSync, rmSync } from "node:fs";
 import path from "path";
+import { cleanFixture } from "./build-helpers";
 import { cli, cliSnapshot } from "./helpers";
 
 const distBin = path.resolve(import.meta.dirname, "../dist/bin/mops.js");
 
-function cleanFixture(cwd: string, ...extras: string[]) {
-  rmSync(path.join(cwd, ".mops"), { recursive: true, force: true });
-  for (const p of extras) {
-    rmSync(p, { recursive: true, force: true });
-  }
-}
-
+// Core build behaviour. The `[optimize]`, Wasm-analysis and PocketIC
+// check-deploy groups live in build-optimize / build-check-wasm /
+// build-check-deploy: jest parallelises across files, not within one, so a
+// single build suite made itself the tail of the entire run.
 describe("build", () => {
   // Several dfx/pocket-ic builds per test; slow CI can exceed 60s default.
   jest.setTimeout(120_000);
@@ -148,310 +140,16 @@ describe("build", () => {
     },
   );
 
-  test("[optimize] runs wasm-opt after build", async () => {
-    const cwd = path.join(import.meta.dirname, "build/optimize");
-    const stamp = path.join(cwd, ".mops/.build/.wasm-opt-ran");
-    try {
-      const result = await cli(["build", "--verbose"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout + result.stderr).toMatch(/Optimized main\.wasm/);
-      expect(existsSync(path.join(cwd, ".mops/.build/main.wasm"))).toBe(true);
-      expect(existsSync(stamp)).toBe(true);
-    } finally {
-      cleanFixture(cwd);
-      rmSync(stamp, { force: true });
-    }
-  });
-
-  test("--no-optimize skips the wasm-opt pass", async () => {
-    const cwd = path.join(import.meta.dirname, "build/optimize");
-    const stamp = path.join(cwd, ".mops/.build/.wasm-opt-ran");
-    try {
-      const result = await cli(["build", "--no-optimize", "--verbose"], {
-        cwd,
-      });
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout + result.stderr).not.toMatch(/Optimized main\.wasm/);
-      expect(existsSync(path.join(cwd, ".mops/.build/main.wasm"))).toBe(true);
-      expect(existsSync(stamp)).toBe(false);
-    } finally {
-      cleanFixture(cwd);
-      rmSync(stamp, { force: true });
-    }
-  });
-
-  test("over-limit IC0505 estimate continues to PocketIC", async () => {
-    const cwd = path.join(import.meta.dirname, "build/wasm-complexity");
-    try {
-      const result = await cli(["build"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toMatch("Warning [MOPS-WASM-COMPLEXITY]");
-      expect(result.stderr).toMatch(
-        /Function: 0[\s\S]*Estimated complexity: 1,000,050[\s\S]*IC0505 limit: 1,000,000[\s\S]*Limit usage: 100\.0%[\s\S]*Instruction count: 20,001/,
-      );
-      expect(result.stderr).toMatch(
-        "Run `mops build --check-deploy` for authoritative PocketIC validation",
-      );
-      expect(result.stderr).toMatch("Error code: CanisterInvalidWasm");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-wasm analyzes Wasm without starting PocketIC", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-wasm-flag");
-    try {
-      const result = await cli(["build", "--check-wasm"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(result.stderr).toMatch("Warning [MOPS-WASM-COMPLEXITY]");
-      expect(result.stdout).not.toMatch("check deploy canister");
-      expect(result.stdout).toMatch("Built 1 canister successfully");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--no-check-deploy leaves configured Wasm analysis enabled", async () => {
-    const cwd = path.join(import.meta.dirname, "build/wasm-complexity");
-    try {
-      const result = await cli(["build", "--no-check-deploy"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(result.stderr).toMatch("Warning [MOPS-WASM-COMPLEXITY]");
-      expect(result.stdout).not.toMatch("check deploy canister");
-      expect(result.stdout).toMatch("Built 1 canister successfully");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--no-check-wasm leaves configured PocketIC validation enabled", async () => {
-    const cwd = path.join(import.meta.dirname, "build/wasm-complexity");
-    try {
-      const result = await cli(["build", "--no-check-wasm"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).not.toMatch("MOPS-WASM-COMPLEXITY");
-      expect(result.stderr).toMatch("Error code: CanisterInvalidWasm");
-      expect(result.stdout).toMatch("check deploy canister main");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("[build].check-deploy installs the built Wasm on PocketIC", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-deploy-config");
-    try {
-      await cliSnapshot(["build"], { cwd }, 0);
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-deploy installs the built Wasm on PocketIC", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-deploy");
-    try {
-      await cliSnapshot(["build", "--check-deploy"], { cwd }, 0);
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-deploy accepts a path-pinned PocketIC binary", async () => {
-    const versionCwd = path.join(import.meta.dirname, "build/check-deploy");
-    const cwd = path.join(import.meta.dirname, "build/check-deploy-path");
-    const localBin = path.join(cwd, "pocket-ic");
-    try {
-      const binResult = await cli(["toolchain", "bin", "pocket-ic"], {
-        cwd: versionCwd,
-      });
-      expect(binResult.exitCode).toBe(0);
-      rmSync(localBin, { force: true });
-      linkSync(binResult.stdout.trim(), localBin);
-
-      const result = await cli(["build", "--check-deploy"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toMatch("check deploy canister main");
-    } finally {
-      rmSync(localBin, { force: true });
-      cleanFixture(cwd);
-    }
-  });
-
-  test("build without check-deploy config does not check deployment", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-deploy");
-    try {
-      const result = await cli(["build"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).not.toMatch("check deploy canister");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--no-check-deploy overrides [build].check-deploy", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-deploy-config");
-    try {
-      const result = await cli(["build", "--no-check-deploy"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).not.toMatch("check deploy canister");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-deploy reports Wasm memory limit failures", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-deploy-fail");
-    try {
-      const result = await cli(["build", "--check-deploy"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toMatch("PocketIC deployment check failed");
-      expect(result.stderr).toMatch("Wasm memory limit");
-      expect(result.stderr).toMatch(
-        "Error code: CanisterWasmMemoryLimitExceeded",
-      );
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-deploy reports Candid encoding errors without a deployment label", async () => {
-    const cwd = path.join(
-      import.meta.dirname,
-      "build/check-deploy-invalid-arg",
-    );
-    try {
-      const result = await cli(["build", "--check-deploy"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toMatch("Invalid initArg for canister main");
-      expect(result.stderr).not.toMatch("PocketIC deployment check failed");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-deploy preserves an ordinary installation failure", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-deploy-trap");
-    try {
-      const result = await cli(["build", "--check-deploy"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toMatch("PocketIC deployment check failed");
-      expect(result.stderr).toMatch("assertion failed");
-      expect(result.stderr).toMatch("Error code: CanisterCalledTrap");
-      expect(result.stderr).not.toMatch("MOPS-CHECK-DEPLOY-SKIPPED");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-deploy preserves PocketIC startup failures", async () => {
-    const cwd = path.join(
-      import.meta.dirname,
-      "build/check-deploy-startup-fail",
-    );
-    try {
-      const result = await cli(["build", "--check-deploy"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toMatch("PocketIC deployment check failed");
-      expect(result.stdout).not.toMatch("check deploy canister");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("check-deploy skips chains incompatible with empty state and checks siblings", async () => {
-    const cwd = path.join(
-      import.meta.dirname,
-      "build/check-deploy-incomplete-migrations",
-    );
-    try {
-      const result = await cli(["build"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(result.stderr).toMatch("Warning [MOPS-CHECK-DEPLOY-SKIPPED]");
-      expect(result.stderr).toMatch("Canister: problematic");
-      expect(result.stderr).toMatch("Canister: problematic2");
-      expect(
-        result.stderr.match(/Fresh PocketIC deployment check did not run\./g),
-      ).toHaveLength(2);
-      expect(result.stderr).toMatch(
-        "moc reported that the generated stable state is incompatible with an empty canister",
-      );
-      expect(result.stderr).toMatch("Compatibility error");
-      expect(result.stdout).not.toMatch("check deploy canister problematic");
-      expect(result.stdout).not.toMatch("check deploy canister problematic2");
-      expect(result.stdout).toMatch("check deploy canister healthy");
-      expect(result.stdout).toMatch("check deploy canister aliased");
-      expect(result.stdout).toMatch("Built 4 canisters successfully");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
+  // Lives here rather than in build-check-deploy because it builds
+  // `build/success`, and this file owns that fixture. Every test in it calls
+  // `cleanFixture`, which removes `.mops` — from a parallel worker that would
+  // delete a build another test is still using.
   test("--check-deploy works without a PocketIC pin (default version)", async () => {
     const cwd = path.join(import.meta.dirname, "build/success");
     try {
       const result = await cli(["build", "foo", "--check-deploy"], { cwd });
       expect(result.exitCode).toBe(0);
       expect(result.stdout).toMatch("check deploy canister foo");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--check-deploy rejects a legacy PocketIC pin", async () => {
-    const cwd = path.join(import.meta.dirname, "build/check-deploy-legacy");
-    const result = await cli(["build", "--check-deploy"], { cwd });
-
-    expect(result.exitCode).toBe(1);
-    expect(result.stderr).toMatch("pins below 9.0.0 no longer work");
-    expect(result.stdout).toMatch("mops toolchain use pocket-ic 14.0.0");
-  });
-
-  test("rejects an invalid wasmMemoryLimit without --check-deploy", async () => {
-    const cwd = path.join(import.meta.dirname, "build/invalid-memory-limit");
-    const result = await cli(["build"], { cwd });
-
-    expect(result.exitCode).toBe(1);
-    expect(result.stderr).toMatch(
-      "Invalid wasmMemoryLimit for canister main: expected a positive integer number of bytes",
-    );
-    expect(result.stderr).not.toMatch("PocketIC deployment check failed");
-  });
-
-  test("[optimize] fails the build when wasm-opt errors", async () => {
-    const cwd = path.join(import.meta.dirname, "build/optimize-fail");
-    try {
-      const result = await cli(["build"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toMatch(/Failed to optimize main\.wasm/);
-      expect(result.stderr).toMatch("mock-wasm-opt: intentional failure");
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("[optimize] without a wasm-opt pin errors instead of pinning one", async () => {
-    const cwd = path.join(import.meta.dirname, "build/optimize-unpinned");
-    const manifest = path.join(cwd, "mops.toml");
-    const before = readFileSync(manifest, "utf8");
-    try {
-      const result = await cli(["build"], { cwd });
-      expect(result.exitCode).toBe(1);
-      expect(result.stderr).toMatch("wasm-opt is not pinned");
-      expect(result.stderr).toMatch("mops toolchain use wasm-opt 131");
-      // The build must not rewrite the manifest it was handed.
-      expect(readFileSync(manifest, "utf8")).toBe(before);
-      // ...and must fail before compiling anything.
-      expect(existsSync(path.join(cwd, ".mops/.build/main.wasm"))).toBe(false);
-    } finally {
-      cleanFixture(cwd);
-    }
-  });
-
-  test("--no-optimize builds despite an unpinned wasm-opt", async () => {
-    const cwd = path.join(import.meta.dirname, "build/optimize-unpinned");
-    try {
-      const result = await cli(["build", "--no-optimize"], { cwd });
-      expect(result.exitCode).toBe(0);
-      expect(existsSync(path.join(cwd, ".mops/.build/main.wasm"))).toBe(true);
     } finally {
       cleanFixture(cwd);
     }
